+++ parallelism/Semaphore.h
@@ -1,31 +1,32 @@
 #ifndef DEF_L_Semaphore
 #define DEF_L_Semaphore
 
-#include "../macros.h"
-
 #if defined L_WINDOWS
-    #include <windows.h>
+#include <windows.h>
 #elif defined L_UNIX
-    #include <semaphore.h>
+#include <semaphore.h>
 #endif
 
-namespace L{
-    class Semaphore{
-        private:
-            #if defined L_WINDOWS
-                HANDLE sem;
-            #elif defined L_UNIX
-                sem_t sem;
-            #endif
-        public:
-            Semaphore(size_t);
-            L_NoCopy(Semaphore)
-            ~Semaphore();
-            void wait();
-            void wait(size_t);
-            void post();
-            void post(size_t);
-    };
+#include "../types.h"
+#include "../macros.h"
+
+namespace L {
+  class Semaphore {
+    private:
+#if defined L_WINDOWS
+      HANDLE sem;
+#elif defined L_UNIX
+      sem_t sem;
+#endif
+    public:
+      Semaphore(uint);
+      L_NoCopy(Semaphore)
+      ~Semaphore();
+      void wait();
+      void wait(uint);
+      void post();
+      void post(uint);
+  };
 }
 
 #endif
